Consolidate duplicate code
