+++ java/app-encryption/src/main/java/com/godaddy/asherah/crypto/engine/bouncycastle/BouncyAes256GcmCrypto.java
@@ -3,17 +3,8 @@ package com.godaddy.asherah.crypto.engine.bouncycastle;
 import org.bouncycastle.crypto.engines.AESEngine;
 import org.bouncycastle.crypto.modes.AEADBlockCipher;
 import org.bouncycastle.crypto.modes.GCMBlockCipher;
-import org.bouncycastle.crypto.params.AEADParameters;
-import org.bouncycastle.crypto.params.KeyParameter;
-
-import com.godaddy.asherah.crypto.keys.CryptoKey;
 
 public class BouncyAes256GcmCrypto extends BouncyAeadCrypto {
-
-  //Do NOT modify the size of the nonce unless you've read NIST 800-38D
-  private static final int NonceSizeBits = 96;
-  private static final int MacSizeBits = 128;
-  private static final int KeySizeBits = 256;
 
   /**
    * Creates a new {@link com.godaddy.asherah.crypto.engine.bouncycastle.BouncyAeadCrypto} instance.
@@ -26,27 +17,4 @@ public class BouncyAes256GcmCrypto extends BouncyAeadCrypto {
   protected AEADBlockCipher getNewAeadBlockCipherInstance() {
     return new GCMBlockCipher(new AESEngine());
   }
-
-  @Override
-  protected AEADParameters getParameters(final CryptoKey key, final byte[] nonce) {
-    return key.withKey(keyBytes -> {
-      return new AEADParameters(new KeyParameter(keyBytes), MacSizeBits, nonce);
-    });
-  }
-
-  @Override
-  protected int getNonceSizeBits() {
-    return NonceSizeBits;
-  }
-
-  @Override
-  protected int getKeySizeBits() {
-    return KeySizeBits;
-  }
-
-  @Override
-  protected int getMacSizeBits() {
-    return MacSizeBits;
-  }
-
 }
