+++ java/app-encryption/src/main/java/com/godaddy/asherah/crypto/engine/bouncycastle/BouncyChaCha20Poly1305Crypto.java
@@ -2,17 +2,8 @@ package com.godaddy.asherah.crypto.engine.bouncycastle;
 
 import org.bouncycastle.crypto.modes.AEADCipher;
 import org.bouncycastle.crypto.modes.ChaCha20Poly1305;
-import org.bouncycastle.crypto.params.AEADParameters;
-import org.bouncycastle.crypto.params.KeyParameter;
-
-import com.godaddy.asherah.crypto.keys.CryptoKey;
 
 public class BouncyChaCha20Poly1305Crypto extends BouncyAeadCrypto {
-
-  //Do NOT modify the size of the nonce unless you've read NIST 800-38D
-  private static final int NonceSizeBits = 96;
-  private static final int MacSizeBits = 128;
-  private static final int KeySizeBits = 256;
 
   /**
    * Creates a new {@link BouncyAeadCrypto} instance.
@@ -25,27 +16,4 @@ public class BouncyChaCha20Poly1305Crypto extends BouncyAeadCrypto {
   protected AEADCipher getNewAeadBlockCipherInstance() {
     return new ChaCha20Poly1305();
   }
-
-  @Override
-  protected AEADParameters getParameters(final CryptoKey key, final byte[] nonce) {
-    return key.withKey(keyBytes -> {
-      return new AEADParameters(new KeyParameter(keyBytes), MacSizeBits, nonce);
-    });
-  }
-
-  @Override
-  protected int getNonceSizeBits() {
-    return NonceSizeBits;
-  }
-
-  @Override
-  protected int getKeySizeBits() {
-    return KeySizeBits;
-  }
-
-  @Override
-  protected int getMacSizeBits() {
-    return MacSizeBits;
-  }
-
 }
